+++ src/types.ts
@@ -17,37 +17,8 @@ export type AccountApiParameters = {
   api_key?: string;
   timeout?: number;
 };
-export type AccountInformation = {
-  account_email: string;
-  account_id: string;
-  account_rate_limit_per_hour: number;
-  api_key: string;
-  extra_credits: number;
-  last_hour_searches: number;
-  plan_id: string;
-  plan_name: string;
-  plan_searches_left: number;
-  searches_per_month: number;
-  this_hour_searches: number;
-  this_month_usage: number;
-  total_searches_left: number;
-};
-
 export type LocationsApiParameters = {
   q?: string;
   limit?: number;
   timeout?: number;
 };
-export type Location = {
-  canonical_name: string;
-  country_code: string;
-  google_id: number;
-  google_parent_id: number;
-  gps: [number, number];
-  id: string;
-  keys: string[];
-  name: string;
-  reach: number;
-  target_type: string;
-};
-export type Locations = Location[];