+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to
 
 ### Removed
 
-- Remove all types for engine parameters. SerpApi's
+- Remove all types for engine parameters and responses. SerpApi's
   [documentation](https://serpapi.com/search-api) should be the only source of
   truth for valid engines and their parameters.